Split main.rs into modules
