+++ crates/punktf-cli/README.md
@@ -1,0 +1,6 @@
+# punktf cli
+
+This is where the command line interface of punktf is located. In particular,
+`main.rs` with the entry point (`fn main`) for the executable.
+
+This crates holds all cli argument definitions together with helper functions.
